Fixed incorrect call of JumpOn action and unify other jumping stuff.
JumpInSub incorrectly calls the text from JumpOn in one case. But more
importantly, JumpIn, JumpOn, and JumpOver all distinguish animate nouns
in the appropriate Sub, but then JumpIn and JumpOn fail to provide the
proper message.

This patch also unifies the handling, since the text for JumpOver was a
little different. In particular, I think the default for JumpOver should
suggest that the action wasn't taken at all, which is a good default for
"jump over <scenery>" or any other thing which is simply too large or
nonsensical.