Stop player from entering a door that goes nowhere.
If an object has the "door" attribute but lacks a "door_dir" property, then
attempting to ENTER or SIT ON the door will cause the player to be sent
inside the door.  This is an abnormal condition and results in the player
being sent somewhere outside of the object tree.  So now a door must also
have the "door_dir" property to be enterable or else the player will be
told off for trying to enter it.