Fix problem with attempting to do a notin test on a routine.
Originally reported problem of fatal error doing the same thing still
needs investigation.

The problem was initially reported at
https://intfiction.org/t/inform-6-12-lib-fatal-error/41261/8 wherein
ASK HER ABOUT STUFF in a stub game would result in a fatal error:

===begin quote===
Welcome!

MYGAME1
My first Inform game.
Release 1 / Serial number 190503 / Inform v6.33 Library 6/12

Your study
There is a doorway to the east of this austere room.

You can see a writing desk here.

>ask her about stuff
@Attempt to address illegal object 7228.  This is normally fatal.
Fatal error: Illegal object
[Hit any key to exit.]
===end quote===

Unfortunately, I could not get this test game nor my own to compile such
that this fatal error is provoked.  Instead, I got a complaint about
doing a "notin" test on a routine:

===begin quote===
>ask her about stuff

[** Programming error: tried to test "in" or "notin" of <routine 7223> **]

[** Programming error: tried to test "in" or "notin" of <routine 7223> **]

[** Programming error: tried to test "in" or "notin" of <routine 7223> **]
I'm not sure what "her" refers to.

>
===end quote===

So the solution for that is to add a check that the noun in question
isn't a routine to the sequence of ANDed tests before trying an
ImplicitTake() on that noun.